requirements.txt
requests==2.26.0
unoffical-athena-federation-sdk==0.0.2
